+++ src/componets/App.js
@@ -5,21 +5,35 @@ import Home from 'routes/Home';
 
 function App() {
     const [init, setInit] = useState(false);
-    const [isLoggedIn, setIsLoggedIn] = useState(false);
     const [userObj, setUserObj] = useState(null);
     useEffect(() => {
         authService.onAuthStateChanged((user) => {
             if (user) {
-                setIsLoggedIn(user);
-                setUserObj(user);
+                // setIsLoggedIn(user);
+                const user = authService.currentUser;
+                setUserObj({
+                    uid: user.uid,
+                    displayName: user.displayName,
+                    updateProfile: (args) => user.updateProfile(args),
+                });
             } else {
-                setIsLoggedIn(false);
+                setUserObj(false);
             }
             setInit(true);
         });
     }, []);
-
-    return <>{init ? <AppRouter isLoggedIn={isLoggedIn} userObj={userObj} /> : 'initializing...'}</>;
+    const refreshUser = () => {
+        setUserObj(authService.currentUser);
+    };
+    return (
+        <>
+            {init ? (
+                <AppRouter isLoggedIn={Boolean(userObj)} userObj={userObj} refreshUser={refreshUser} />
+            ) : (
+                'initializing...'
+            )}
+        </>
+    );
 }
 
 export default App;
