+++ src/componets/Router.js
@@ -5,14 +5,14 @@ import Auth from '../routes/Auth';
 import Home from '../routes/Home';
 import Navigation from './Navigation';
 
-const AppRouter = ({ isLoggedIn, userObj }) => {
+const AppRouter = ({ isLoggedIn, userObj, refreshUser }) => {
     return (
         <Router>
             <Routes>
-                {isLoggedIn && <Route path="/" element={<Navigation />} />}
+                {isLoggedIn && <Route path="/" element={<Navigation userObj={userObj} />} />}
                 {isLoggedIn ? <Route path="/" element={<Home />} /> : <Route path="/" element={<Auth />} />}
                 <Route path="/home" element={<Home userObj={userObj} />} />
-                <Route path="/profile" element={<Profile />} />
+                <Route path="/profile" element={<Profile userObj={userObj} refreshUser={refreshUser} />} />
             </Routes>
         </Router>
     );
