--- src/routes/EditProfile.js
@@ -1,4 +1,0 @@
-import React from 'react';
-
-const EditProfile = () => <span>EditProfile</span>;
-export default EditProfile;
